put `Connect` implementation inside `storage` package and adjust `README.md`
