+++ storage/README.md
@@ -21,12 +21,8 @@ One interface to connect many of database, This package is aimed not only used f
 
 import "github.com/codenoid/file.io/storage"
 
-stg = storage.Storage{
-	Type: "redis",
-}
-
-// host, username, password, database
-stg.Connect("127.0.0.1:6379", "", "", "0")
+// type://user:pass@host:port/database
+stg = storage.Connect("redis://127.0.0.1:6379/0")
 
 // then you can call this function
 stg.Set()
